requirements.txt
langchain 
PyPDF2 
python-dotenv 
streamlit 
openai 
faiss-cpu 
altair 
tiktoken 
langchain_community 
langchain_openai 